+++ mlflow/tracing/__init__.py
@@ -1,4 +1,5 @@
 from mlflow.tracing.config import configure
+from mlflow.tracing.context import context
 from mlflow.tracing.databricks import set_databricks_monitoring_sql_warehouse_id
 from mlflow.tracing.display import disable_notebook_display, enable_notebook_display
 from mlflow.tracing.distributed import (
@@ -8,15 +9,6 @@ from mlflow.tracing.distributed import (
 from mlflow.tracing.enablement import set_experiment_trace_location, unset_experiment_trace_location
 from mlflow.tracing.provider import disable, enable, reset, set_destination
 from mlflow.tracing.utils import set_span_chat_tools
-
-
-def __getattr__(name):
-    if name == "context":
-        from mlflow.tracing.fluent import context
-
-        return context
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
-
 
 __all__ = [
     "configure",
